Add WellKnown to ApiClient
